server/app/tools/web_search.py
import logging
import json
from typing import Dict, Any, List, Optional
import requests
import time
from bs4 import BeautifulSoup

from app.core.config import settings

logger = logging.getLogger(__name__)

def web_search(query: str, num_results: int = 5) -> List[Dict[str, Any]]:
    """Perform a web search for the given query.
    
    Args:
        query: The search query
        num_results: Number of results to return (default: 5)
    
    Returns:
        List of search result objects with title, snippet, and URL
    """
    # Ensure number of results is within limits
    num_results = min(num_results, settings.WEB_SEARCH_MAX_RESULTS)
    
    # In a production environment, you would integrate with a real search API
    # such as Google Custom Search API, Bing Search API, or similar
    # This is a simulated search result for development/demonstration
    
    logger.info(f"Web search: {query}, results: {num_results}")
    
    try:
        # Simulated search delay to mimic real API call
        time.sleep(0.5)
        
        # Create simulated results for financial searches
        if "stock" in query.lower() or "market" in query.lower() or "financ" in query.lower():
            results = [
                {
                    "title": "Latest Market News and Analysis - Financial Times",
                    "snippet": "Comprehensive financial market news and analysis, covering stocks, bonds, currencies, and more.",
                    "url": "https://www.ft.com/markets"
                },
                {
                    "title": "Yahoo Finance - Stock Market Live, Quotes, News",
                    "snippet": "Real-time stock quotes, financial news, portfolio management tools, and more.",
                    "url": "https://finance.yahoo.com/"
                },
                {
                    "title": "CNBC - Stock Markets, Business News, Investment Analysis",
                    "snippet": "Latest business news, market data, and investment analysis for investors and financial professionals.",
                    "url": "https://www.cnbc.com/markets/"
                },
                {
                    "title": "Bloomberg - Financial News, Stock Quotes, Analysis",
                    "snippet": "Breaking news and in-depth analysis of business, financial markets, and global economies.",
                    "url": "https://www.bloomberg.com/markets"
                },
                {
                    "title": "Investopedia - Financial Analysis and Investment Education",
                    "snippet": "Educational content explaining financial concepts, investment strategies, and market mechanics.",
                    "url": "https://www.investopedia.com/"
                },
            ]
            
            # If the query includes a specific ticker symbol, add stock-specific results
            import re
            ticker_match = re.search(r'\b[A-Z]{1,5}\b', query)
            if ticker_match:
                ticker = ticker_match.group(0)
                ticker_results = [
                    {
                        "title": f"{ticker} Stock Price and Overview - Yahoo Finance",
                        "snippet": f"Detailed stock information for {ticker}, including real-time price quotes, financial data, and company news.",
                        "url": f"https://finance.yahoo.com/quote/{ticker}"
                    },
                    {
                        "title": f"{ticker} Company Profile - MarketWatch",
                        "snippet": f"Company profile, financial statements, and analyst ratings for {ticker}.",
                        "url": f"https://www.marketwatch.com/investing/stock/{ticker}"
                    }
                ]
                # Insert ticker-specific results at the top
                results = ticker_results + results
        else:
            # Generic search results for non-financial queries
            results = [
                {
                    "title": "General search result 1 for: " + query,
                    "snippet": "Snippet description for the first search result that matches your query.",
                    "url": "https://example.com/result1"
                },
                {
                    "title": "General search result 2 for: " + query,
                    "snippet": "Snippet description for the second search result that matches your query.",
                    "url": "https://example.com/result2"
                },
                {
                    "title": "General search result 3 for: " + query,
                    "snippet": "Snippet description for the third search result that matches your query.",
                    "url": "https://example.com/result3"
                }
            ]
        
        # Return the requested number of results
        return results[:num_results]
    
    except Exception as e:
        logger.error(f"Error in web search: {e}", exc_info=True)
        return [
            {
                "title": "Error performing search",
                "snippet": f"An error occurred: {str(e)}",
                "url": None
            }
        ]

def fetch_webpage_content(url: str) -> Dict[str, Any]:
    """Fetch content from a webpage.
    
    Args:
        url: The URL to fetch
    
    Returns:
        Dictionary with title, text content, and status
    """
    try:
        # In a production environment, consider:
        # 1. Using a proper User-Agent header
        # 2. Respecting robots.txt
        # 3. Rate limiting requests
        # 4. Handling cookies and redirects appropriately
        headers = {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
        }
        
        logger.info(f"Fetching webpage: {url}")
        
        response = requests.get(url, headers=headers, timeout=10)
        response.raise_for_status()
        
        # Parse HTML content
        soup = BeautifulSoup(response.text, 'html.parser')
        
        # Get title
        title = soup.title.string if soup.title else "No title found"
        
        # Get main content (simplified)
        # This is a very simple extraction that won't work well for many sites
        # Consider using a proper content extraction library in production
        paragraphs = soup.find_all('p')
        text_content = "\n".join(p.get_text() for p in paragraphs)
        
        return {
            "title": title,
            "content": text_content[:5000],  # Limit content length
            "url": url,
            "status": "success"
        }
    
    except Exception as e:
        logger.error(f"Error fetching webpage {url}: {e}", exc_info=True)
        return {
            "title": "Error",
            "content": f"Failed to fetch content: {str(e)}",
            "url": url,
            "status": "error"
        } 