server/requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]

# Configuration & Environment
python-dotenv
pydantic
pydantic-settings

# Agent Frameworks
langchain
langgraph
langchain-openai # Or langchain-anthropic, langchain-google-vertexai, etc.
langchain-community # For various integrations
langchain-text-splitters

# Multi-Agent Communication
autogenstudio # Includes pyautogen

# Context Retrieval (RAG)
llama-index
llama-index-llms-openai # Or other llm integrations
llama-index-embeddings-openai # Or other embedding integrations
llama-index-vector-stores-chroma # Example vector store, adjust as needed (e.g., pgvector)
llama-index-readers-file

# Data Layer (PostgreSQL + SQLAlchemy Async)
database_url # For parsing DATABASE_URL
sqlalchemy[asyncio]
asyncpg # Async driver for PostgreSQL
alembic # For database migrations (optional but recommended)
psycopg2-binary # Required by Alembic/SQLAlchemy even for async sometimes
greenlet # Required by SQLAlchemy

# Fintech Libraries
pandas
pandas-ta # For Technical Indicators
# TA-Lib # Uncomment if needed - Requires manual installation of TA-Lib C library first!
vectorbt # For Backtesting
# backtrader # Alternative backtesting library
# zipline-reloaded # Alternative backtesting library (may have complex dependencies)
skopt # For scikit-optimize (Bayesian Optimization)
# optuna # Alternative hyperparameter optimization library
scikit-learn # For Grid Search and other ML utilities
numpy

# Web Scraping/Requests
requests
beautifulsoup4
# playwright # Uncomment if advanced scraping/browser automation is needed

# Utilities
python-multipart # For FastAPI forms/uploads
passlib[bcrypt] # For password hashing
python-jose[cryptography] # For JWT

# Logging
loguru # Example flexible logging library
